+++ app/src/main/java/com/studio4plus/homerplayer/model/Position.java
@@ -2,9 +2,9 @@ package com.studio4plus.homerplayer.model;
 
 public class Position {
     public final String filePath;
-    public final int seekPosition;
+    public final long seekPosition;
 
-    public Position(String filePath, int seekPosition) {
+    public Position(String filePath, long seekPosition) {
         this.filePath = filePath;
         this.seekPosition = seekPosition;
     }
